........
  r65532 | georg.brandl | 2008-08-04 22:35:07 +0000 (Mon, 04 Aug 2008) | 2 lines

  Add a test file for misc. markup and fix a doc bug.
........
  r65546 | georg.brandl | 2008-08-05 09:55:20 +0000 (Tue, 05 Aug 2008) | 2 lines

  Show an indication while downloading the search index.
........
  r65547 | georg.brandl | 2008-08-05 10:07:50 +0000 (Tue, 05 Aug 2008) | 2 lines

  Fix an indentation problem in production lists. #3477.
........
  r65550 | georg.brandl | 2008-08-05 16:21:34 +0000 (Tue, 05 Aug 2008) | 2 lines

  Remove unused interface.js.
........
  r65551 | georg.brandl | 2008-08-06 08:48:21 +0000 (Wed, 06 Aug 2008) | 2 lines

  Don't try to remove a nonexisting static dir.
........

+++ sphinx/builder.py
@@ -604,7 +604,8 @@ class StandaloneHTMLBuilder(Builder):
                 if path.isfile(fullname):
                     shutil.copyfile(fullname, targetname)
                 elif path.isdir(fullname):
-                    shutil.rmtree(targetname)
+                    if path.exists(targetname):
+                        shutil.rmtree(targetname)
                     shutil.copytree(fullname, targetname)
         # copy logo file (handled differently)
         if self.config.html_logo:
